Final-project/requirements.txt
streamlit==1.34.0
beautifulsoup4==4.12.3
requests==2.31.0
numpy==1.26.4
sentence-transformers==2.7.0
google-generativeai==0.3.2
supabase==2.3.1
pandas==2.2.1
python-dotenv==1.0.1
googlesearch-python==1.2.3
urllib3==2.2.1
serpapi==1.3.0